Squashed commit of the following:
commit de19f34e5eab0b5b35ac022478d0eb2e4e74fd1c
Author: Shiyo <strikinginsanity@hotmail.com>
Date:   Wed Sep 17 12:27:50 2025 -0500

    Remove healing from WS scripts

commit 6de54f8ab6dae33a3f49d22e9d1c60cc6fe2aed9
Author: Shiyo <strikinginsanity@hotmail.com>
Date:   Wed Sep 17 12:26:56 2025 -0500

    Retaliation mod fix

commit 7f6b4230f56736110623a8146826a3c84dcc7414
Author: Shiyo <strikinginsanity@hotmail.com>
Date:   Wed Sep 17 12:26:50 2025 -0500

    (WS) Add an Action packet for healing WS

commit 9bda2e820e93a2f03f17abf9ae67f797650b7e33
Author: Shiyo <strikinginsanity@hotmail.com>
Date:   Tue Sep 16 18:53:01 2025 -0500

    TODO: Only shows up on target and 1 secondary target

+++ src/map/utils/battleutils.cpp
@@ -1388,7 +1388,7 @@ namespace battleutils
                                                               dmg, PHYSICAL_ATTACK_TYPE::NORMAL, SLOT_MAIN);
 
                     // Apply Retaliation bonus modifier
-                    uint16 bonus = dmg * (PDefender->getMod(Mod::RETALIATION) / 100);
+                    uint16 bonus = static_cast<uint16>(dmg * (PDefender->getMod(Mod::RETALIATION) / 100.0));
                     dmg += bonus;
 
                     // Apply the damage and handle TP and stoneskin
